Move all messages in to one place and clean up messages and log messages

+++ src/main/java/dev/emortal/velocity/permissions/commands/subs/user/UserRoleRemoveSub.java
@@ -6,13 +6,11 @@ import dev.emortal.api.command.CommandExecutor;
 import dev.emortal.api.service.permission.PermissionService;
 import dev.emortal.api.service.permission.RemoveRoleFromPlayerResult;
 import dev.emortal.api.utils.resolvers.PlayerResolver;
+import dev.emortal.velocity.lang.ChatMessages;
 import dev.emortal.velocity.permissions.PermissionCache;
-import dev.emortal.velocity.permissions.commands.subs.role.RoleSubUtils;
-import io.grpc.Status;
 import io.grpc.StatusException;
 import io.grpc.StatusRuntimeException;
-import net.kyori.adventure.text.minimessage.MiniMessage;
-import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
+import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,14 +19,6 @@ import java.util.UUID;
 
 public final class UserRoleRemoveSub implements CommandExecutor<CommandSource> {
     private static final Logger LOGGER = LoggerFactory.getLogger(UserRoleRemoveSub.class);
-    private static final MiniMessage MINI_MESSAGE = MiniMessage.miniMessage();
-
-    private static final String ROLE_NOT_FOUND = "<red>Role <role_id> not found";
-    private static final String ROLE_REMOVED = "<green>Role <role_id> removed from user <username>";
-    private static final String ALREADY_HAS_ROLE = "<red>User <user_id> already has the <role_id> role";
-    private static final String DOESNT_HAVE_ROLE = "<red>User <user_id> doesn't have role <role_id>";
-
-    private static final String PERMISSION_PLAYER_NOT_FOUND = "<red>Player <uuid> not found in permission service";
 
     private final PermissionService permissionService;
     private final PermissionCache permissionCache;
@@ -44,41 +34,42 @@ public final class UserRoleRemoveSub implements CommandExecutor<CommandSource> {
         String targetUsername = context.getArgument("username", String.class);
         String roleId = context.getArgument("roleId", String.class);
 
-        PermissionCache.CachedRole role = RoleSubUtils.getRole(this.permissionCache, context);
-        if (role == null) return;
-
-        PlayerResolver.CachedMcPlayer playerData;
-        try {
-            playerData = PlayerResolver.getPlayerData(targetUsername);
-        } catch (StatusException exception) {
-            Status status = exception.getStatus();
-            var usernamePlaceholder = Placeholder.unparsed("username", targetUsername);
-            if (status.getCode() == Status.Code.NOT_FOUND) {
-                source.sendMessage(MINI_MESSAGE.deserialize("<red>Player <username> not found", usernamePlaceholder));
-            } else {
-                source.sendMessage(MINI_MESSAGE.deserialize("<red>Failed to retrieve player data for <username>", usernamePlaceholder));
-            }
+        PermissionCache.CachedRole role = this.permissionCache.getRole(roleId);
+        if (role == null) {
+            ChatMessages.ERROR_ROLE_NOT_FOUND.send(source, Component.text(roleId));
             return;
         }
 
-        UUID targetId = playerData.uuid();
-        String correctUsername = playerData.username();
+        PlayerResolver.CachedMcPlayer target;
+        try {
+            target = PlayerResolver.getPlayerData(targetUsername);
+        } catch (StatusException exception) {
+            LOGGER.error("Failed to get player data for '{}'", targetUsername, exception);
+            ChatMessages.GENERIC_ERROR.send(source);
+            return;
+        }
+
+        if (target == null) {
+            ChatMessages.PLAYER_NOT_FOUND.send(source, Component.text(targetUsername));
+            return;
+        }
+
+        UUID targetId = target.uuid();
+        String correctUsername = target.username();
 
         RemoveRoleFromPlayerResult result;
         try {
             result = this.permissionService.removeRoleFromPlayer(targetId, roleId);
         } catch (StatusRuntimeException exception) {
-            LOGGER.error("Something went wrong removing role from user", exception);
-            source.sendMessage(MINI_MESSAGE.deserialize("<red>Something went wrong"));
+            LOGGER.error("Failed to remove role '{}' from '{}'", roleId, targetUsername, exception);
+            ChatMessages.GENERIC_ERROR.send(source);
             return;
         }
 
-        var roleIdPlaceholder = Placeholder.unparsed("role_id", roleId);
-        var message = switch (result) {
-            case SUCCESS -> MINI_MESSAGE.deserialize(ROLE_REMOVED, roleIdPlaceholder, Placeholder.unparsed("username", correctUsername));
-            case DOES_NOT_HAVE_ROLE -> MINI_MESSAGE.deserialize(DOESNT_HAVE_ROLE, roleIdPlaceholder);
-            case PLAYER_NOT_FOUND -> MINI_MESSAGE.deserialize(PERMISSION_PLAYER_NOT_FOUND, Placeholder.unparsed("uuid", targetId.toString()));
-        };
-        source.sendMessage(message);
+        switch (result) {
+            case SUCCESS -> ChatMessages.USER_ROLE_REMOVED.send(source, Component.text(roleId), Component.text(correctUsername));
+            case DOES_NOT_HAVE_ROLE -> ChatMessages.ERROR_USER_MISSING_ROLE.send(source, Component.text(roleId));
+            case PLAYER_NOT_FOUND -> ChatMessages.PLAYER_NOT_FOUND.send(source, Component.text(correctUsername));
+        }
     }
 }
